git-svn-id: http://svn.code.sf.net/p/netpbm/code/advanced@1256 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ converter/other/exif.c
@@ -818,8 +818,8 @@ process_EXIF(unsigned char * const ExifData,
             
         if (ShowTags){
             fprintf(stderr, 
-                    "Non-settings part of Exif header: %d bytes\n",
-                    ExifData+length-LastExifRefd);
+                    "Non-settings part of Exif header: %lu bytes\n",
+                    (unsigned long)(ExifData+length-LastExifRefd));
         }
     }
 }
